+++ src/pages/planetPages/marsPage/MarsPage.js
@@ -1,13 +1,17 @@
 import React from "react";
-import { 
-  Button, 
-  Paragraf, 
-  Headder, 
-  Seksjon, 
-  Kilde 
+
+import {
+  Button,
+  Paragraf,
+  Headder,
+  Seksjon,
+  Kilde,
+  Flex,
+  Text,
+  Liste,
 } from "./styles";
 
-const MarsPage = () => {
+const marsPage = () => {
   return (
     <Seksjon>
       <div className='container mars__container'>
@@ -15,14 +19,48 @@ const MarsPage = () => {
           <Headder>Mars</Headder>
 
           <Paragraf>
-            Mars er den fjerde planeten regnet fra Sola. Den er den syvende
-            største planeten i solsystemet og har litt over halvparten så stor
-            omkrets som Jorda. Den er i likhet med Merkur, Venus og Jorda en
-            steinplanet. Et døgn på Mars varer 24,6 timer, ikke ulikt et døgn på
-            Jorda (23,9 timer). Et Mars-døgn kalles sol, som er en forkortelse
-            for «solar day» (soldøgn). Mars kan ha vært mer jordlignende for 3,5
-            milliarder år siden. Formasjoner på overflaten tyder på at vann har
-            rent der tidligere, mens i dag er overflaten tørrlagt.{" "}
+            <Flex>
+              <Text>
+                <p>
+                  Mars er den fjerde planeten regnet fra Sola. Den er den
+                  syvende største planeten i solsystemet og har litt over
+                  halvparten så stor omkrets som Jorda. Den er i likhet med
+                  Merkur, Venus og Jorda en steinplanet.
+                  <br /> Et døgn på Mars varer 24,6 timer, ikke ulikt et døgn på
+                  Jorda (23,9 timer). <br />
+                  Et Mars-døgn kalles sol, som er en forkortelse for «solar day»
+                  (soldøgn).
+                  <br /> Mars kan ha vært mer jordlignende for 3,5 milliarder år
+                  siden. Formasjoner på overflaten tyder på at vann har rent der
+                  tidligere, mens i dag er overflaten tørrlagt.{" "}
+                </p>
+              </Text>
+              <p>
+                <Liste>
+                  <h4>Fakta om Mars</h4>
+                  <ul>
+                    <li>
+                      Gjennomsnittlig avstand fra Sola 227,9 millioner kilometer
+                    </li>
+                    <li>Banens eksentrisitet 0,093</li>
+                    <li>
+                      Banens helning i forhold til ekliptikken 1,85 grader
+                    </li>
+                    <li>Radius ved ekvator (Jorda = 1) 0,532</li>
+                    <li>Masse (Jorda = 1) 0,107</li>
+                    <li>Tyngdeakselerasjon på overflaten 3,71 m/s2</li>
+                    <li>
+                      Gjennomsnittlig overflatetemperatur cirka −55 grader
+                      celsius
+                    </li>
+                    <li>Siderisk omløpstid 1,881 år</li>
+                    <li>Rotasjonstid 24,623 timer</li>
+                    <li>Måner 2</li>
+                  </ul>
+                </Liste>
+              </p>
+            </Flex>
+
             <Kilde>
               Aksnes, Kaare; Hammerstrøm, Maria: Mars i Store norske leksikon på
               snl.no. Hentet 11. oktober 2022 fra
@@ -35,4 +73,4 @@ const MarsPage = () => {
   );
 };
 
-export default MarsPage;
+export default marsPage;
