+++ src/pages/planetPages/moonPage/styles.js
@@ -1,64 +1,46 @@
 import styled from "styled-components";
-import Moon from "./img/moon.jpg";
+import MoonImage from "./img/earthFromMoon.jpg";
 
-export const Seksjon = styled.section`
-  background-image: url(${Moon});
-  background-repeat: no-repeat;
-  background-size: cover;
-  width: 100%; /*or your image's width*/
-  height: 100vh; /*or your image's height*/
+export const Box = styled.div`
+  margin: 10rem 7rem;
   @media (max-width: 768px) {
-    width: 86%;
-  }
-  @media (max-width: 375px) {
-    width: 90%;
+    margin: 7rem 1rem;
   }
 `;
 
-export const Headder = styled.h1`
-  color: #ffff;
-  font-size: 64px;
+export const MoonH1 = styled.h1`
+  font-size: 5rem;
+  font-weight: bold;
+  text-align: center;
+  -webkit-text-stroke: 2px black;
 `;
 
-export const Bilde = styled.img`
-  width: 50px;
-  height: 50px;
+export const MoonP = styled.p`
+  font-size: 2rem;
+  font-weight: bold;
+  -webkit-text-stroke: 1px black;
+  text-align: left;
+  padding-right: 20%;
+  margin-bottom: 2rem;
+  @media (max-width: 768px) {
+    font-size: 1rem;
+    padding: 0;
+  }
 `;
 
-export const Paragraf = styled.p`
-  color: #ffff;
-  font-size: 32px;
+export const MoonContainer = styled.div`
+  position: fixed;
+  bottom: 0rem;
+  left: 0rem;
+  width: 100%;
+  height: 100%;
+  background: url(${MoonImage}) no-repeat;
+  background-size: cover;
+  z-index: -1;
 `;
 
-export const Flex = styled.p`
-  display: flex;
-  flex-direction: row;
-  flex-wrap: nowrap;
-  gap: 80px;
-  justify-content: space-around;
-  align-items: stretch;
-  align-content: space-around;
-`;
-
-export const Liste = styled.p`
-  color: #ffff;
-  font-size: 16px;
-  margin: 50px 60px 30px 50px;
-`;
-
-export const Text = styled.p`
-  color: #ffff;
-  font-size: 20px;
-  margin: 100px 60px 30px 0px;
-`;
-
-export const Kilde = styled.span`
-  color: white;
-  font-size: 16px;
-`;
-
-export const Button = styled.a`
-  padding-left: 5px;
-  background-color: transparent;
-  color: white;
+export const MoonImg = styled.img`
+  height: 600%;
+  width: 200%;
+  object-fit: contain;
 `;
